refactor GestureMenu-related component to be under GestureMenu
