+++ src/components/AppComponent.tsx
@@ -1,7 +1,6 @@
 import { Capacitor } from '@capacitor/core'
 import { StatusBar, Style } from '@capacitor/status-bar'
 import _ from 'lodash'
-import { animate, useMotionValue } from 'motion/react'
 import React, { FC, PropsWithChildren, useEffect, useLayoutEffect, useRef } from 'react'
 import { useSelector } from 'react-redux'
 import { WebviewBackground } from 'webview-background'
@@ -27,12 +26,12 @@ import DesktopCommandUniverse from './DesktopCommandUniverse'
 import DropGutter from './DropGutter'
 import ErrorMessage from './ErrorMessage'
 import Footer from './Footer'
-import GestureMenu from './GestureMenu'
+import GestureContentBlur from './GestureMenu/GestureContentBlur'
+import GestureMenu from './GestureMenu/GestureMenu'
 import HamburgerMenu from './HamburgerMenu'
 import LatestCommandsDiagram from './LatestCommandsDiagram'
 import MultiGesture from './MultiGesture'
 import NavBar from './NavBar'
-import ProgressiveBlur from './ProgressiveBlur'
 import Sidebar from './Sidebar'
 import Tips from './Tips/Tips'
 import Toolbar from './Toolbar'
@@ -113,60 +112,6 @@ const MultiGestureIfTouch: FC<PropsWithChildren> = ({ children }) => {
     </MultiGesture>
   ) : (
     <>{children}</>
-  )
-}
-
-/**
- * Renders a progressive blur over the app content while the gesture menu is open.
- *
- * Composed as a sibling of <Content/> so it lands inside MultiGesture's react-native-web <View> at
- * runtime. Per CSS painting order the blur (position:absolute, z-index:auto) paints above the in-flow
- * Content but below the position:fixed; z-index:24 gesture trace — so the content behind the menu
- * blurs while the trace stays sharp. An overlay rendered outside View cannot achieve this because
- * View's z-index:0 stacking context traps the trace below any root-level element.
- * See docs/superpowers/specs/2026-06-26-gesture-trace-above-blur-root-cause-and-decision.md.
- */
-const GestureContentBlur: FC = () => {
-  const animationState = gestureStore.useSelector(state => state.gestureMenuAnimationState)
-  const blurHeight = gestureStore.useSelector(state => state.gestureMenuBlurHeight)
-
-  // A MotionValue drives the opacity of each backdrop-filter layer individually. Animating opacity on a
-  // shared parent of backdrop-filter elements breaks the blur in WebKit, so ProgressiveBlur applies the
-  // MotionValue per-layer rather than via a wrapping element.
-  const blurOpacity = useMotionValue(0)
-  useEffect(() => {
-    const controls = animate(blurOpacity, animationState === 'visible' ? 1 : 0, { duration: 0.15, ease: 'easeOut' })
-    return controls.stop
-  }, [animationState, blurOpacity])
-
-  // Content (position:relative; zIndex:content) is positioned, so it paints in the positive-z-index
-  // phase — above a z-index:auto overlay. The blur must therefore carry an explicit z-index that sits
-  // above content but below the gesture trace (gestureTrace) for its backdrop-filter to soften the
-  // content while leaving the trace sharp. All three share the gesture <View>'s stacking context.
-  return (
-    <div
-      className={css({
-        position: 'absolute',
-        top: 0,
-        left: 0,
-        right: 0,
-        zIndex: 'gestureContentBlur',
-        pointerEvents: 'none',
-      })}
-      // Cap the wrapper to the menu footprint (rem string from gestureStore); since ProgressiveBlur is
-      // position:absolute; inset:0 it fills the wrapper, so this caps the blur + feather to the menu height.
-      style={{ height: blurHeight }}
-    >
-      <ProgressiveBlur
-        // Max blur at the top, fading to 0 downward — matches the menu's top-weighted falloff.
-        direction='to bottom'
-        maxBlur={8}
-        opacity={blurOpacity}
-        // Feather the bottom edge so the blur fades out instead of cutting off.
-        mask='linear-gradient(180deg, black 0%, black 80%, transparent 100%)'
-        // Sizing the blur to the menu footprint is handled by the companion height-strategy spec.
-      />
-    </div>
   )
 }
