+++ src/components/MultiGesture.tsx
@@ -1,5 +1,5 @@
 import React, { PropsWithChildren } from 'react'
-import { GestureResponderEvent, PanResponder, PanResponderInstance, View } from 'react-native'
+import { GestureResponderEvent, PanResponder, PanResponderInstance, View, ViewStyle } from 'react-native'
 import Direction from '../@types/Direction'
 import Gesture from '../@types/Gesture'
 import { noop } from '../constants'
@@ -345,7 +345,12 @@ class MultiGesture extends React.Component<MultiGestureProps> {
   render() {
     const ref = React.createRef<HTMLDivElement>()
     return (
-      <View {...this.panResponder.panHandlers}>
+      <View
+        {...this.panResponder.panHandlers}
+        // View's default z-index:0 traps children below NavBar's stacking context; z-index:auto
+        // removes it, letting gesture blur/trace layer above NavBar in the root context.
+        style={{ zIndex: 'auto' } as unknown as ViewStyle}
+      >
         <TraceGesture eventNodeRef={ref} />
         <ScrollZone leftHanded={this.leftHanded} />
         <div ref={ref}>{this.props.children}</div>
